feat(#21): moved types
* types now in another file
