+++ src/builder/fileBatch.ts
@@ -2,10 +2,8 @@ import fs from "node:fs";
 import path from "node:path";
 import { NAME_FORMAT } from "../utils/constants.js";
 import { HL7FatalError, HL7ParserError } from "../utils/exception.js";
-import {
-  ClientBuilderFileOptions,
-  normalizedClientFileBuilderOptions,
-} from "../utils/normalizedBuilder.js";
+import { normalizedClientFileBuilderOptions } from "../utils/normalizedBuilder.js";
+import { ClientBuilderFileOptions } from "../utils/types.js";
 import { createHL7Date, split } from "../utils/utils.js";
 import { Batch } from "./batch.js";
 import { HL7Node } from "./interface/hL7Node.js";
