+++ src/builder/modules/rootBase.ts
@@ -1,6 +1,6 @@
 import { Delimiters } from "../../utils/enum.js";
 import { HL7FatalError } from "../../utils/exception.js";
-import { ClientBuilderOptions } from "../../utils/normalizedBuilder.js";
+import { ClientBuilderOptions } from "../../utils/types.js";
 import { decodeHexString, escapeForRegExp } from "../../utils/utils.js";
 import { NodeBase } from "./nodeBase.js";
